+++ lib/src/gesture/page_view/widgets/page_controller.dart
@@ -1,8 +1,8 @@
 // ignore_for_file: unnecessary_null_comparison, always_put_control_body_on_new_line
 
-import 'dart:math' as math;
-import 'package:flutter/foundation.dart';
-import 'package:flutter/material.dart';
+import 'package:flutter/widgets.dart';
+
+import 'scroll_position.dart';
 
 class ExtendedPageController extends ScrollController {
   /// Creates a page controller.
@@ -13,10 +13,38 @@ class ExtendedPageController extends ScrollController {
     this.keepPage = true,
     this.viewportFraction = 1.0,
     this.pageSpacing = 0.0,
+    this.shouldIgnorePointerWhenScrolling = false,
   })  : assert(initialPage != null),
         assert(keepPage != null),
         assert(viewportFraction != null),
         assert(viewportFraction > 0.0);
+
+  /// Whether the contents of the widget should ignore [PointerEvent] inputs.
+  ///
+  /// Setting this value to true prevents the use from interacting with the
+  /// contents of the widget with pointer events. The widget itself is still
+  /// interactive.
+  ///
+  /// For example, if the scroll position is being driven by an animation, it
+  /// might be appropriate to set this value to ignore pointer events to
+  /// prevent the user from accidentally interacting with the contents of the
+  /// widget as it animates. The user will still be able to touch the widget,
+  /// potentially stopping the animation.
+  ///
+  ///
+  /// if true, we should handle scale event in [ExtendedImageGesturePageView] before [ExtendedImageGesturePageView] stop scroll.
+  /// notice: there is one issue that we may be zoom two image at the same time, because we can't find out which one should be zoomed.
+  ///
+  ///
+  /// if false, Image can accept scale event before [ExtendedImageGesturePageView] stop scroll.
+  /// notice: we don't know  there are any issues if we don't ignore [PointerEvent] inputs when it's scrolling.
+  ///
+  ///
+  /// Two way to solve issue that we can's zoom image before [PageView] stop scroll.
+  ///
+  ///
+  /// default is false.
+  final bool shouldIgnorePointerWhenScrolling;
 
   /// The page to show when first creating the [PageView].
   final int initialPage;
@@ -74,7 +102,7 @@ class ExtendedPageController extends ScrollController {
       'The page property cannot be read when multiple PageViews are attached to '
       'the same PageController.',
     );
-    final _PagePosition position = this.position as _PagePosition;
+    final ExtendedPagePosition position = this.position as ExtendedPagePosition;
     return position.page;
   }
 
@@ -89,7 +117,7 @@ class ExtendedPageController extends ScrollController {
     required Duration duration,
     required Curve curve,
   }) {
-    final _PagePosition position = this.position as _PagePosition;
+    final ExtendedPagePosition position = this.position as ExtendedPagePosition;
     return position.animateTo(
       position.getPixelsFromPage(page.toDouble()),
       duration: duration,
@@ -102,7 +130,7 @@ class ExtendedPageController extends ScrollController {
   /// Jumps the page position from its current value to the given value,
   /// without animation, and without checking if the new value is in range.
   void jumpToPage(int page) {
-    final _PagePosition position = this.position as _PagePosition;
+    final ExtendedPagePosition position = this.position as ExtendedPagePosition;
     position.jumpTo(position.getPixelsFromPage(page.toDouble()));
   }
 
@@ -130,7 +158,7 @@ class ExtendedPageController extends ScrollController {
   @override
   ScrollPosition createScrollPosition(ScrollPhysics physics,
       ScrollContext context, ScrollPosition? oldPosition) {
-    return _PagePosition(
+    return ExtendedPagePosition(
       physics: physics,
       context: context,
       initialPage: initialPage,
@@ -138,211 +166,15 @@ class ExtendedPageController extends ScrollController {
       viewportFraction: viewportFraction,
       oldPosition: oldPosition,
       pageSpacing: pageSpacing,
+      shouldIgnorePointerWhenScrolling: shouldIgnorePointerWhenScrolling,
     );
   }
 
   @override
   void attach(ScrollPosition position) {
     super.attach(position);
-    final _PagePosition pagePosition = position as _PagePosition;
+    final ExtendedPagePosition pagePosition = position as ExtendedPagePosition;
     pagePosition.viewportFraction = viewportFraction;
-  }
-}
-
-class _PagePosition extends ScrollPositionWithSingleContext
-    implements PageMetrics {
-  _PagePosition({
-    required ScrollPhysics physics,
-    required ScrollContext context,
-    this.initialPage = 0,
-    bool keepPage = true,
-    double viewportFraction = 1.0,
-    ScrollPosition? oldPosition,
-    double pageSpacing = 0.0,
-  })  : assert(initialPage != null),
-        assert(keepPage != null),
-        assert(viewportFraction != null),
-        assert(viewportFraction > 0.0),
-        assert(pageSpacing != null),
-        assert(pageSpacing >= 0.0),
-        _viewportFraction = viewportFraction,
-        _pageSpacing = pageSpacing,
-        _pageToUseOnStartup = initialPage.toDouble(),
-        super(
-          physics: physics,
-          context: context,
-          initialPixels: null,
-          keepScrollOffset: keepPage,
-          oldPosition: oldPosition,
-        );
-
-  final int initialPage;
-  double _pageToUseOnStartup;
-
-  @override
-  Future<void> ensureVisible(
-    RenderObject object, {
-    double alignment = 0.0,
-    Duration duration = Duration.zero,
-    Curve curve = Curves.ease,
-    ScrollPositionAlignmentPolicy alignmentPolicy =
-        ScrollPositionAlignmentPolicy.explicit,
-    RenderObject? targetRenderObject,
-  }) {
-    // Since the _PagePosition is intended to cover the available space within
-    // its viewport, stop trying to move the target render object to the center
-    // - otherwise, could end up changing which page is visible and moving the
-    // targetRenderObject out of the viewport.
-    return super.ensureVisible(
-      object,
-      alignment: alignment,
-      duration: duration,
-      curve: curve,
-      alignmentPolicy: alignmentPolicy,
-      targetRenderObject: null,
-    );
-  }
-
-  @override
-  double get viewportFraction => _viewportFraction;
-  double _viewportFraction;
-  set viewportFraction(double value) {
-    if (_viewportFraction == value) return;
-    final double? oldPage = page;
-    _viewportFraction = value;
-    if (oldPage != null) forcePixels(getPixelsFromPage(oldPage));
-  }
-
-  double _pageSpacing;
-  double get pageSpacing => _pageSpacing;
-  set pageSpacing(double value) {
-    if (_pageSpacing != value) {
-      final double? oldPage = page;
-      _pageSpacing = value;
-      if (oldPage != null) forcePixels(getPixelsFromPage(oldPage));
-    }
-  }
-
-  // fix viewportDimension
-  @override
-  double get viewportDimension => super.viewportDimension + pageSpacing;
-
-  // The amount of offset that will be added to [minScrollExtent] and subtracted
-  // from [maxScrollExtent], such that every page will properly snap to the center
-  // of the viewport when viewportFraction is greater than 1.
-  //
-  // The value is 0 if viewportFraction is less than or equal to 1, larger than 0
-  // otherwise.
-  double get _initialPageOffset =>
-      math.max(0, viewportDimension * (viewportFraction - 1) / 2);
-
-  double getPageFromPixels(double pixels, double viewportDimension) {
-    final double actual = math.max(0.0, pixels - _initialPageOffset) /
-        math.max(1.0, viewportDimension * viewportFraction);
-    final double round = actual.roundToDouble();
-    if ((actual - round).abs() < precisionErrorTolerance) {
-      return round;
-    }
-    return actual;
-  }
-
-  double getPixelsFromPage(double page) {
-    return page * viewportDimension * viewportFraction + _initialPageOffset;
-  }
-
-  @override
-  double? get page {
-    assert(
-      !hasPixels || (minScrollExtent != null && maxScrollExtent != null),
-      'Page value is only available after content dimensions are established.',
-    );
-    return !hasPixels
-        ? null
-        : getPageFromPixels(
-            pixels.clamp(minScrollExtent, maxScrollExtent), viewportDimension);
-  }
-
-  @override
-  void saveScrollOffset() {
-    PageStorage.of(context.storageContext)?.writeState(
-        context.storageContext, getPageFromPixels(pixels, viewportDimension));
-  }
-
-  @override
-  void restoreScrollOffset() {
-    if (!hasPixels) {
-      final double? value = PageStorage.of(context.storageContext)
-          ?.readState(context.storageContext) as double?;
-      if (value != null) _pageToUseOnStartup = value;
-    }
-  }
-
-  @override
-  void saveOffset() {
-    context.saveOffset(getPageFromPixels(pixels, viewportDimension));
-  }
-
-  @override
-  void restoreOffset(double offset, {bool initialRestore = false}) {
-    assert(initialRestore != null);
-    assert(offset != null);
-    if (initialRestore) {
-      _pageToUseOnStartup = offset;
-    } else {
-      jumpTo(getPixelsFromPage(offset));
-    }
-  }
-
-  @override
-  bool applyViewportDimension(double viewportDimension) {
-    final double? oldViewportDimensions =
-        hasViewportDimension ? this.viewportDimension : null;
-    // fix viewportDimension
-    if (viewportDimension + pageSpacing == oldViewportDimensions) {
-      return true;
-    }
-    final bool result = super.applyViewportDimension(viewportDimension);
-    final double? oldPixels = hasPixels ? pixels : null;
-    final double page = (oldPixels == null || oldViewportDimensions == 0.0)
-        ? _pageToUseOnStartup
-        : getPageFromPixels(oldPixels, oldViewportDimensions!);
-    final double newPixels = getPixelsFromPage(page);
-
-    if (newPixels != oldPixels) {
-      correctPixels(newPixels);
-      return false;
-    }
-    return result;
-  }
-
-  @override
-  bool applyContentDimensions(double minScrollExtent, double maxScrollExtent) {
-    final double newMinScrollExtent = minScrollExtent + _initialPageOffset;
-    return super.applyContentDimensions(
-      newMinScrollExtent,
-      math.max(newMinScrollExtent, maxScrollExtent - _initialPageOffset),
-    );
-  }
-
-  @override
-  PageMetrics copyWith({
-    double? minScrollExtent,
-    double? maxScrollExtent,
-    double? pixels,
-    double? viewportDimension,
-    AxisDirection? axisDirection,
-    double? viewportFraction,
-  }) {
-    return PageMetrics(
-      minScrollExtent: minScrollExtent ??
-          (hasContentDimensions ? this.minScrollExtent : null),
-      maxScrollExtent: maxScrollExtent ??
-          (hasContentDimensions ? this.maxScrollExtent : null),
-      pixels: pixels ?? (hasPixels ? this.pixels : null),
-      viewportDimension: viewportDimension ??
-          (hasViewportDimension ? this.viewportDimension : null),
-      axisDirection: axisDirection ?? this.axisDirection,
-      viewportFraction: viewportFraction ?? this.viewportFraction,
-    );
+    pagePosition.pageSpacing = pageSpacing;
   }
 }